fully replace slow e_000 by fast e_000
